+++ src/SimpleRingBuffer.hpp
@@ -8,14 +8,14 @@
 template <typename T>
 class SimpleRingBuffer {
   public:
-    SimpleRingBuffer(size_t _buffSize) : buffSize(_buffSize), writeMode(0), deallocateBuffer(true) {
+    SimpleRingBuffer(uint32_t _buffSize) : buffSize(_buffSize), writeMode(0), deallocateBuffer(true) {
       buffer = new T[buffSize];
       if(!buffer)
         Serial.printf("failed buffer = new T[%d]\n", buffSize);
       init();
     }
 
-    SimpleRingBuffer(T *_buffer, size_t _buffSize) : buffSize(_buffSize), buffer(_buffer), writeMode(0), deallocateBuffer(false) {
+    SimpleRingBuffer(T *_buffer, uint32_t _buffSize) : buffSize(_buffSize), buffer(_buffer), writeMode(0), deallocateBuffer(false) {
       init();
     }
 
@@ -31,22 +31,22 @@ class SimpleRingBuffer {
       writeSum = 0;
     }
     
-    size_t available() {
+    uint32_t available() {
       if (readPtr > writePtr)
         return buffSize + writePtr - readPtr;
       else
         return            writePtr - readPtr;
     }
     
-    size_t size() {
+    uint32_t size() {
       return buffSize;
     }
     
-    size_t free() {
+    uint32_t free() {
       return buffSize - available() - 1;
     }
     
-    size_t write(T *payload, size_t length) {
+    uint32_t write(T *payload, uint32_t length) {
       if (!payload || !length || length > buffSize) return 0;
 
       auto f = free();
@@ -61,7 +61,7 @@ class SimpleRingBuffer {
         }
       }
 
-      auto bytesAvail = std::min(length, (size_t)(buffSize - writePtr));
+      auto bytesAvail = std::min(length, (uint32_t)(buffSize - writePtr));
       memcpy(&buffer[writePtr], &payload[0         ], sizeof(T) * bytesAvail);
       memcpy(&buffer[0       ], &payload[bytesAvail], sizeof(T) * (length - bytesAvail));
       writePtr = (writePtr + length) % buffSize;
@@ -70,7 +70,7 @@ class SimpleRingBuffer {
       return length;
     }
 
-    size_t write(std::function<size_t(T * payload, size_t length)> func, size_t length = UINT16_MAX) {     
+    uint32_t write(std::function<uint32_t(T * payload, uint32_t length)> func, uint32_t length = UINT16_MAX) {     
 /*    if(!length)
         length = f;
       else if (length > f) {
@@ -90,8 +90,8 @@ class SimpleRingBuffer {
       length = std::min(length, f);
       if(!length) return 0;
       
-      size_t wlen = 0;
-      auto bytesAvail = std::min(length, (size_t)(buffSize - writePtr));
+      uint32_t wlen = 0;
+      auto bytesAvail = std::min(length, (uint32_t)(buffSize - writePtr));
       if(bytesAvail > 0)
         wlen = func(&buffer[writePtr], bytesAvail);
 
@@ -104,11 +104,11 @@ class SimpleRingBuffer {
       return wlen;
     }
 
-    size_t read(T *data, size_t length) {
+    uint32_t read(T *data, uint32_t length) {
       length = std::min(length, available());
       if (!data || !length) return 0;
 
-      auto bytesAvail = std::min(length, (size_t)(buffSize - readPtr));
+      auto bytesAvail = std::min(length, (uint32_t)(buffSize - readPtr));
       memcpy(&data[0         ], &buffer[readPtr], sizeof(T) * bytesAvail);
       memcpy(&data[bytesAvail], &buffer[0      ], sizeof(T) * (length - bytesAvail));
       readPtr = (readPtr + length) % buffSize;
@@ -117,8 +117,8 @@ class SimpleRingBuffer {
       return length;
     }
 
-    size_t seek(size_t length) {
-      size_t result = std::min( available(), length );
+    uint32_t seek(uint32_t length) {
+      uint32_t result = std::min( available(), length );
       readPtr = (readPtr + length) % buffSize;
       return result;
     }
@@ -128,8 +128,8 @@ class SimpleRingBuffer {
         + " Free:" + String(free()) 
         + " WritePtr:" + String(writePtr) 
         + " ReadPtr:" + String(readPtr) 
-        + " WriteSum:" + ( writeSum > (uint64_t)UINT32_MAX ? String((size_t)(writeSum / 1024)) + "K" : String((size_t)writeSum) )
-        + " readSum:"  + ( readSum  > (uint64_t)UINT32_MAX ? String((size_t)( readSum / 1024)) + "K" : String((size_t) readSum) );
+        + " WriteSum:" + ( writeSum > (uint64_t)UINT32_MAX ? String((uint32_t)(writeSum / 1024)) + "K" : String((uint32_t)writeSum) )
+        + " readSum:"  + ( readSum  > (uint64_t)UINT32_MAX ? String((uint32_t)( readSum / 1024)) + "K" : String((uint32_t) readSum) );
     }
     
     uint64_t getTotalRead() {
@@ -142,9 +142,9 @@ class SimpleRingBuffer {
     
   private:
     T *buffer;
-    size_t buffSize;
-    size_t readPtr;
-    size_t writePtr;
+    uint32_t buffSize;
+    uint32_t readPtr;
+    uint32_t writePtr;
     uint8_t writeMode;     // 空きが足りないときの挙動 0:writeしない 1:空きの分だけ 2:先頭を削除して強制挿入
     bool deallocateBuffer;
     uint64_t writeSum;
