+++ src/WebRadio_Jcbasimul.hpp
@@ -9,7 +9,7 @@
 #include "AudioFileSourceJcbasimul.hpp"
 #include "WebRadio.h"
 
-static constexpr const size_t decode_stack_size = 16*1024;
+static constexpr const uint32_t decode_stack_size = 16*1024;
 static constexpr const char* station_list[][2] =
 {
 /*00*/	{"ＦＭはな / 北海道","fmhana"},
@@ -152,18 +152,18 @@ static byte initial_station = 96;   // ウメダFM Be Happy!789
 class Jcbasimul : public WebRadio {
   public:
 #ifndef SEPARATE_DOWNLOAD_TASK
-    Jcbasimul(AudioOutput * _out, int cpuDecode, const UBaseType_t priorityDecode = 3, const size_t buffSize = 0) : bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, priorityDecode) {
+    Jcbasimul(AudioOutput * _out, int cpuDecode, const UBaseType_t priorityDecode = 3, const uint32_t buffSize = 0) : bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, priorityDecode) {
 #else
-    Jcbasimul(AudioOutput * _out, int cpuDecode, const UBaseType_t priorityDecode = 3, const size_t buffSize = 0) : bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, priorityDecode, 1 - cpuDecode, 4096) {
+    Jcbasimul(AudioOutput * _out, int cpuDecode, const UBaseType_t priorityDecode = 3, const uint32_t buffSize = 0) : bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, priorityDecode, 1 - cpuDecode, 4096) {
 #endif
       for(int i = 0; i < sizeof(station_list) / sizeof(station_list[0]); i++)
         stations.push_back(new station_t(this, station_list[i][0], station_list[i][1]));
       defaultStationIdx = initial_station;
     }
 #ifndef SEPARATE_DOWNLOAD_TASK
-    Jcbasimul(AudioOutput * _out, int cpuDecode, uint8_t * _buffer, const size_t buffSize) : buffer(_buffer), bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, 3) {
+    Jcbasimul(AudioOutput * _out, int cpuDecode, uint8_t * _buffer, const uint32_t buffSize) : buffer(_buffer), bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, 3) {
 #else
-    Jcbasimul(AudioOutput * _out, int cpuDecode, uint8_t * _buffer, const size_t buffSize) : buffer(_buffer), bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, 3, 1 - cpuDecode, 4096) {
+    Jcbasimul(AudioOutput * _out, int cpuDecode, uint8_t * _buffer, const uint32_t buffSize) : buffer(_buffer), bufferSize(buffSize), WebRadio(_out, cpuDecode, decode_stack_size, 3, 1 - cpuDecode, 4096) {
 #endif
       for(int i = 0; i < sizeof(station_list) / sizeof(station_list[0]); i++)
         stations.push_back(new station_t(this, station_list[i][0], station_list[i][1]));
@@ -240,7 +240,7 @@ class Jcbasimul : public WebRadio {
       }
       
       if(!bufferSize)
-        bufferSize = std::max(5 * 1024, (int)std::min( (size_t)(128 * 1024), psram));
+        bufferSize = std::max(5UL * 1024UL, std::min(128UL * 1024UL, (uint32_t)psram));
       
       startTask();
       return true; 
@@ -380,7 +380,7 @@ class Jcbasimul : public WebRadio {
     AudioFileSourceJcbasimul * source = nullptr;
     AudioGeneratorOpus * decoder = nullptr;
     uint8_t * buffer = nullptr;
-    size_t bufferSize;
+    uint32_t bufferSize;
     byte stopDecode = 0;    // 2:request stop 1:accept request
 
 };
